requirements.txt
Flask==0.12.1
isodate==0.5.4
python-dateutil==2.6.0
git+git://github.com/bgschiller/winnow.git@0.0.2
